Split vmaas e2e into sanity/regression buckets
Mirrors BMaaS pattern (OSAC-4389): path-based split into sanity (PR
gating, ~10min) vs regression (nightly schedule, ~30min). Sanity bucket
= 7 files with 1 VM boot + auth/networking/catalog API smoke;
regression = all metering, storage-tier matrix, console, external-ip,
negative/validation tests.

Changes:
- Move 7 sanity files to tests/e2e/vmaas/sanity/
- Move 18 regression files + external_ip/ to tests/e2e/vmaas/regression/
- Workflow: PR/merge_group → vmaas/sanity; schedule → vmaas/regression;
  dispatch defaults to sanity, full vmaas opt-in
- Import fixes: external_ip package references updated for new path
- conftest.py stays at vmaas root, shared by both buckets

Estimated PR pytest time reduction: ~35min → ~10-12min.
