+++ tests/e2e/vmaas/regression/__init__.py
@@ -1,0 +1,1 @@
+# VMaaS E2E regression tests — comprehensive coverage for scheduled runs
