+++ tests/e2e/vmaas/regression/test_compute_instance_delete_during_provision.py
