+++ tests/e2e/vmaas/regression/test_compute_instance_disk_image.py
