+++ tests/e2e/vmaas/regression/test_compute_instance_instance_type.py
