+++ tests/e2e/vmaas/sanity/__init__.py
@@ -1,0 +1,1 @@
+# VMaaS E2E sanity tests — fast smoke coverage for PR gating
