+++ tests/e2e/vmaas/sanity/test_compute_instance_cli_fields.py
