+++ tests/e2e/vmaas/sanity/test_disk_image_lifecycle.py
