+++ tests/e2e/vmaas/sanity/test_instance_type_lifecycle.py
