+++ tests/e2e/vmaas/sanity/test_virtual_network_lifecycle.py
